Do not use persistent status sections for status messages
Initially the plugin used `window.status_message`s which disappear
after a few seconds.  A persistent message like "sublack: reformatted"
which doesn't even go away when editing the file is not helpful.

Persistent messages for the daemon *on the view* are also a mistake
as they stick to the view although the daemon has been probably started
for the window.  Make them non-sticky as well.